fix: refactor some - extract control btns
